Revert "add translation for Main"
This reverts commit 28f7e6f3b604ef63f70f6aeba11998151a8f607a.
