--- .storybook/preview.tsx
@@ -1,58 +1,0 @@
-import { Preview } from '@storybook/react'
-
-// .storybook/preview.js
-import React, { Suspense, useEffect } from 'react'
-import { I18nextProvider } from 'react-i18next'
-import i18n from '../src/i18n'
-
-// Wrap your stories in the I18nextProvider component
-const withI18next = (Story, context) => {
-  const { locale } = context.globals
-
-  // When the locale global changes
-  // Set the new locale in i18n
-  useEffect(() => {
-    i18n.changeLanguage(locale)
-  }, [locale])
-
-  return (
-    // This catches the suspense from components not yet ready (still loading translations)
-    // Alternative: set useSuspense to false on i18next.options.react when initializing i18next
-    <Suspense fallback={<div>loading translations...</div>}>
-      <I18nextProvider i18n={i18n}>
-        <Story />
-      </I18nextProvider>
-    </Suspense>
-  )
-}
-
-// export decorators for storybook to wrap your stories in
-export const decorators = [withI18next]
-
-const preview: Preview = {
-  parameters: {
-    actions: { argTypesRegex: '^on[A-Z].*' },
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/,
-      },
-    },
-  },
-  globalTypes: {
-    locale: {
-      name: 'Lang',
-      description: 'Internationalization locale',
-      toolbar: {
-        icon: 'globe',
-        items: [
-          { value: 'en', title: 'English' },
-          { value: 'ru', title: 'Russian' },
-        ],
-        showName: true,
-      },
-    },
-  },
-}
-
-export default preview
