+++ package.json
@@ -9,8 +9,8 @@
     "lint": "next lint",
     "test": "jest",
     "test:w": "jest --watch",
-    "sb": "storybook dev -p 6006 -s ./public",
-    "build-storybook": "storybook build -s public",
+    "sb": "storybook dev -p 6006",
+    "build-storybook": "storybook build",
     "cr:com": "generate-react component",
     "cr:ui": "generate-react component --type=ui",
     "cr:lay": "generate-react component --type=layouts"
@@ -23,7 +23,6 @@
     "eslint": "8.38.0",
     "eslint-config-next": "13.3.0",
     "i18next": "^22.4.14",
-    "i18next-http-backend": "^2.2.0",
     "next": "13.3.0",
     "next-i18next": "^13.2.2",
     "react": "18.2.0",
@@ -41,12 +40,10 @@
     "@storybook/testing-library": "^0.0.14-next.2",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
-    "@types/i18next-browser-languagedetector": "^3.0.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-storybook": "^0.6.11",
     "generate-react-cli": "^8.3.0",
     "husky": "^8.0.3",
-    "i18next-browser-languagedetector": "^7.0.1",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
     "sass": "^1.62.0",
