+++ public/locales/en/buttons.json
@@ -9,4 +9,5 @@
   "subscription": "Subscription for 99 ₽",
   "plug": "Connect for 99 ₽",
   "activateCertificate": "Activate certificate"
+
 }
