+++ src/ui/SelectedButton/SelectedButton.stories.tsx
@@ -1,3 +1,4 @@
+
 import { Meta, StoryObj } from '@storybook/react'
 
 import { SelectorButton } from './SelectorButton'
